+++ Libraries/Xtro.MDX/FunctionsErrorHandling.h
@@ -11,7 +11,7 @@ public:
 		return gcnew String(DXGetErrorStringW(Result));
 	}
 
-	static int Trace(String^ File, unsigned long Line, int Result, String^ Message, bool PopMessageBox)
+	static int Trace(String^ File, unsigned int Line, int Result, String^ Message, bool PopMessageBox)
 	{
 		IntPtr pFile = IntPtr::Zero;
 		IntPtr pMessage = IntPtr::Zero;
@@ -32,32 +32,32 @@ public:
 	}
 
 #if defined(DEBUG) | defined(_DEBUG)
-	static int TraceMessage(String^ File, unsigned long Line, String^ String)
+	static int TraceMessage(String^ File, unsigned int Line, String^ String)
 	{
 		return Trace(File, Line, 0, String, false);
 	}
 
-	static int TraceError(String^ File, unsigned long Line, String^ String, int Result)
+	static int TraceError(String^ File, unsigned int Line, String^ String, int Result)
 	{
 		return Trace(File, Line, Result, String, false);
 	}
 
-	static int TraceErrorMessageBox(String^ File, unsigned long Line, String^ String, int Result)
+	static int TraceErrorMessageBox(String^ File, unsigned int Line, String^ String, int Result)
 	{
 		return Trace(File, Line, Result, String, true);
 	}
 #else
-	static int TraceMessage(String^ File, unsigned long Line, String^ String)
+	static int TraceMessage(String^ File, unsigned int Line, String^ String)
 	{
 		return 0;
 	}
 
-	static int TraceError(String^ File, unsigned long Line, String^ String, int Result)
+	static int TraceError(String^ File, unsigned int Line, String^ String, int Result)
 	{
 		return Result;
 	}
 
-	static int TraceErrorMessageBox(String^ File, unsigned long Line, String^ String, int Result)
+	static int TraceErrorMessageBox(String^ File, unsigned int Line, String^ String, int Result)
 	{
 		return Result;
 	}
